importの循環参照防止のためtypedDictの定義を別ファイルに分離
